# Conflicts:
#	features/merchant/variant_common/src/main/java/com/tokopedia/variant_common/view/holder/VariantChipViewHolder.kt

+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/createreputation/model/ProductRevResponse.kt
@@ -9,7 +9,7 @@ data class ProductRevResponse(
         val productData: ProductData = ProductData(),
         @SerializedName("reputationID")
         @Expose
-        val reputationID: Int = 0,
+        val reputationID: Long = 0,
         @SerializedName("orderID")
         @Expose
         val orderID: String = "",
